Add support for "voiced" phrases.
The title music sequence has an extra two bytes in the phrase struct
which seem to be related to the APU registers for the pulse channels.
This change lets you construct such patterns although the pitch and
duration enumerations do not work for them.  Instead, you must provide
the raw pattern bytes.

In the future, I will add support for semantic understanding of the
title music note data.