Fix bug with phrase table offsets in sequence data.
The sequence data points to the phrase table structs, but the routine to
create the data had a hardcoded phrase struct size of 6.  Title music
phrase structs are 8 bytes long.  To keep things generic, the routine
now calculates the offsets by checking the length of each phrase struct
instead of assuming all are 6 bytes long.