Transpose palace music so boss music works.
The two pieces were 5/7 semitones apart and it sounded weird when it
transitioned.  This makes them 3/9 semitones apart which works much
better.  I wanted to have them be in the same key but the pitch LUT
doesn't cooperate with that idea and I haven't come up with a good way
to rewrite it yet.